requirements.txt
asgiref
Django
django-cors-headers
django-environ
django-filter
djangorestframework
djangorestframework_simplejwt
drf-yasg
inflection
packaging
pillow
PyJWT
pytz
PyYAML
sqlparse
tzdata
uritemplate
